refactor: move query translation logic to QueryTranslation component
- Move generateTranslation function from SearchView to QueryTranslation
- Remove translation state and logic from SearchView
- Make QueryTranslation self-contained with its own translation logic
- Keep code DRY by centralizing translation logic in one place
- QueryTranslation now handles all query expansion and author resolution

This improves separation of concerns and makes the code more maintainable.
