+++ src/components/QueryTranslation.tsx
@@ -1,15 +1,159 @@
 'use client';
 
-import { useState } from 'react';
+import { useState, useEffect, useCallback, useRef } from 'react';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { faEquals, faChevronDown, faChevronUp } from '@fortawesome/free-solid-svg-icons';
+import { expandParenthesizedOr, parseOrQuery } from '@/lib/search';
+import { resolveAuthorToNpub } from '@/lib/vertex';
+import { applySimpleReplacements } from '@/lib/search/replacements';
+import { nip19 } from 'nostr-tools';
 
 interface QueryTranslationProps {
-  translation: string;
+  query: string;
 }
 
-export default function QueryTranslation({ translation }: QueryTranslationProps) {
+export default function QueryTranslation({ query }: QueryTranslationProps) {
   const [isExplanationExpanded, setIsExplanationExpanded] = useState(false);
+  const [translation, setTranslation] = useState<string>('');
+  const authorResolutionCache = useRef<Map<string, string>>(new Map());
+
+  const generateTranslation = useCallback(async (query: string, skipAuthorResolution = false): Promise<string> => {
+    try {
+      // 1) Apply simple replacements first
+      const afterReplacements = await applySimpleReplacements(query);
+
+      // 2) Recursive OR substitution (distribute parentheses)
+      const distributed = expandParenthesizedOr(afterReplacements);
+
+      // Helper: resolve all by:<author> tokens within a single query string
+      const resolveByTokensInQuery = async (q: string): Promise<string> => {
+        const rx = /(^|\s)by:(\S+)/gi;
+        let result = '';
+        let lastIndex = 0;
+        let m: RegExpExecArray | null;
+        while ((m = rx.exec(q)) !== null) {
+          const full = m[0];
+          const pre = m[1] || '';
+          const raw = m[2] || '';
+          const match = raw.match(/^([^),.;]+)([),.;]*)$/);
+          const core = (match && match[1]) || raw;
+          const suffix = (match && match[2]) || '';
+          let replacement = core;
+          
+          if (!skipAuthorResolution && !/^npub1[0-9a-z]+$/i.test(core)) {
+            // Check cache first
+            if (authorResolutionCache.current.has(core)) {
+              replacement = authorResolutionCache.current.get(core) || core;
+            } else {
+              try {
+                const npub = await resolveAuthorToNpub(core);
+                if (npub) {
+                  replacement = npub;
+                  authorResolutionCache.current.set(core, npub);
+                }
+              } catch {}
+            }
+          }
+          result += q.slice(lastIndex, m.index);
+          result += `${pre}by:${replacement}${suffix}`;
+          lastIndex = m.index + full.length;
+        }
+        result += q.slice(lastIndex);
+        return result;
+      };
+
+      // Helper: normalize p:<token> where token may be hex, npub or nprofile
+      const resolvePTokensInQuery = (q: string): string => {
+        const rx = /(^|\s)p:(\S+)/gi;
+        let result = '';
+        let lastIndex = 0;
+        let m: RegExpExecArray | null;
+        while ((m = rx.exec(q)) !== null) {
+          const full = m[0];
+          const pre = m[1] || '';
+          const raw = m[2] || '';
+          const match = raw.match(/^([^),.;]+)([),.;]*)$/);
+          const core = (match && match[1]) || raw;
+          const suffix = (match && match[2]) || '';
+          let replacement = core;
+          if (/^[0-9a-fA-F]{64}$/.test(core)) {
+            try { replacement = nip19.npubEncode(core.toLowerCase()); } catch {}
+          } else if (/^npub1[0-9a-z]+$/i.test(core)) {
+            replacement = core;
+          } else if (/^nprofile1[0-9a-z]+$/i.test(core)) {
+            try {
+              const decoded = nip19.decode(core);
+              if (decoded?.type === 'nprofile') {
+                const pk = (decoded.data as { pubkey: string }).pubkey;
+                replacement = nip19.npubEncode(pk);
+              }
+            } catch {}
+          }
+          result += q.slice(lastIndex, m.index);
+          result += `${pre}p:${replacement}${suffix}`;
+          lastIndex = m.index + full.length;
+        }
+        result += q.slice(lastIndex);
+        return result;
+      };
+
+      // 3) Resolve authors inside each distributed branch (if not skipping)
+      const resolvedDistributed = skipAuthorResolution 
+        ? distributed 
+        : await Promise.all(distributed.map((q) => resolveByTokensInQuery(q)));
+
+      const withPResolved = resolvedDistributed.map((q) => resolvePTokensInQuery(q));
+
+      // 4) For parenthesized OR expansion, show all expanded queries
+      // Don't split further if we already have multiple distributed queries
+      if (distributed.length > 1) {
+        // We have parenthesized OR expansion - show all expanded queries
+        const preview = withPResolved.join('\n');
+        return preview;
+      }
+
+      // 5) Split into multiple queries if top-level OR exists (for non-parenthesized OR)
+      const finalQueriesSet = new Set<string>();
+      for (const q of withPResolved) {
+        const parts = parseOrQuery(q);
+        if (parts.length > 1) {
+          parts.forEach((p) => { const s = p.trim(); if (s) finalQueriesSet.add(s); });
+        } else {
+          const s = q.trim(); if (s) finalQueriesSet.add(s);
+        }
+      }
+      const finalQueries = Array.from(finalQueriesSet);
+
+      // Format compact preview
+      const preview = finalQueries.length > 0 ? finalQueries.join('\n') : afterReplacements;
+      return preview;
+    } catch {
+      return '';
+    }
+  }, [authorResolutionCache]);
+
+  // Generate translation when query changes
+  useEffect(() => {
+    let cancelled = false;
+    
+    if (!query.trim()) {
+      setTranslation('');
+      return;
+    }
+
+    const generateAndSetTranslation = async () => {
+      const result = await generateTranslation(query, false);
+      if (!cancelled) {
+        setTranslation(result);
+      }
+    };
+
+    generateAndSetTranslation();
+
+    return () => { 
+      cancelled = true; 
+    };
+  }, [query, generateTranslation]);
 
   if (!translation) return null;
